+++ packages/data-schema/src/SchemaProcessor.ts
@@ -399,7 +399,7 @@ function calculateAuth(authorization: Authorization<any, any, any>[]) {
 }
 
 function capitalize<T extends string>(s: T): Capitalize<T> {
-  return `${s[0].toUpperCase()}${s.slice(1)}` as any;
+  return `${s[0].toUpperCase()}${s.slice(1)}` as Capitalize<T>;
 }
 
 function uncapitalize<T extends string>(s: T): Uncapitalize<T> {
@@ -654,6 +654,23 @@ type TransformedSecondaryIndexes = {
 };
 
 /**
+ *
+ * @param pk - partition key field name
+ * @param sk - (optional) array of sort key field names
+ * @returns default query field name
+ */
+const secondaryIndexDefaultQueryField = (
+  pk: string,
+  sk?: readonly string[],
+): string => {
+  const skName = sk?.length ? 'And' + sk?.map(capitalize).join('And') : '';
+
+  const queryField = `listBy${capitalize(pk)}${skName}`;
+
+  return queryField;
+};
+
+/**
  * Given InternalModelIndexType[] returns a map where the key is the model field to be annotated with an @index directive
  * and the value is an array of transformed Amplify @index directives with all supplied attributes
  */
@@ -661,12 +678,15 @@ const transformedSecondaryIndexesForModel = (
   secondaryIndexes: readonly InternalModelIndexType[],
 ): TransformedSecondaryIndexes => {
   const indexDirectiveWithAttributes = (
+    partitionKey: string,
     sortKeys: readonly string[],
     indexName: string,
     queryField: string,
   ): string => {
     if (!sortKeys.length && !indexName && !queryField) {
-      return '@index';
+      return `@index(queryField: "${secondaryIndexDefaultQueryField(
+        partitionKey,
+      )}")`;
     }
 
     const attributes: string[] = [];
@@ -683,6 +703,13 @@ const transformedSecondaryIndexesForModel = (
 
     if (queryField) {
       attributes.push(`queryField: "${queryField}"`);
+    } else {
+      attributes.push(
+        `queryField: "${secondaryIndexDefaultQueryField(
+          partitionKey,
+          sortKeys,
+        )}"`,
+      );
     }
 
     return `@index(${attributes.join(', ')})`;
@@ -696,6 +723,7 @@ const transformedSecondaryIndexesForModel = (
       acc[partitionKey] = acc[partitionKey] || [];
       acc[partitionKey].push(
         indexDirectiveWithAttributes(
+          partitionKey,
           sortKeys as readonly string[],
           indexName,
           queryField,